requirements.txt
streamlit==1.32.2
onnxruntime==1.17.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
torch==2.2.1
Pillow==10.2.0
